requirements.txt
# LLM & Agent Framework
langchain>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
google-generativeai>=0.8.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0

# Validation
pydantic>=2.0.0

# Environment & Config
python-dotenv>=1.0.0

# Audit Log
# (SQLite is built-in to Python)

# UI
streamlit>=1.38.0


# Utilities
tabulate>=0.9.0
